% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agep2.R
\name{agep2}
\alias{agep2}
\title{agep2}
\usage{
agep2(
  betas,
  method = c("horvath", "hannum", "phenoage", "skinblood", "lin", "all"),
  n_missing = TRUE,
  missing_probes = FALSE
)
}
\arguments{
\item{betas}{Matrix of betas}

\item{method}{Default is "horvath"}

\item{n_missing}{Logical, additionally output the number of missing CpGs
for each sample using the specified method or coeff list.}

\item{missing_probes}{Logical, additionally output the names of missing CpGs
for each sample using the specified method or coeff list.}
}
\value{
Returns matrix of predicted ages per sample.
}
\description{
agep2
}
