% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateCellCounts2_GMset.R
\name{estimateCellCounts2_GMset}
\alias{estimateCellCounts2_GMset}
\title{estimateCellCounts2_GMset}
\source{
https://bioconductor.org/packages/FlowSorted.Blood.EPIC/
}
\usage{
estimateCellCounts2_GMset(
  GenomicMethylSet,
  referenceset = NULL,
  CustomCpGs = NULL,
  returnAll = FALSE,
  meanPlot = FALSE,
  verbose = TRUE,
  lessThanOne = FALSE,
  cellcounts = NULL
)
}
\arguments{
\item{GenomicMethylSet}{GenomicMethylSet}

\item{referenceset}{referenceset}

\item{CustomCpGs}{CustomCpGs}

\item{returnAll}{returnAll}

\item{meanPlot}{meanPlot}

\item{verbose}{verbose}

\item{lessThanOne}{lessThanOne}

\item{cellcounts}{cellcounts}
}
\value{
cell proportion
}
\description{
This function is adapted from the \code{estimateCellCounts2} function
in the FlowSorted.Blood.EPIC package.
The original copyright belongs to the FlowSorted.Blood.EPIC authors
and the package is licensed under GPL-3.
}
\author{
Original authors of FlowSorted.Blood.EPIC
}
