% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methxy.dmp.R
\name{methxy.dmp}
\alias{methxy.dmp}
\title{methxy.dmp}
\usage{
methxy.dmp(matrix, Design.matrix, use_eBayes = F, adjust.method = "bonferroni")
}
\arguments{
\item{matrix}{beta or M matrix.}

\item{Design.matrix}{Some users need to adjust for covariates,
so here only the design matrix is input.}

\item{use_eBayes}{default F; use eBayes}

\item{adjust.method}{"bonferroni" or "BH". default "bonferroni"}
}
\value{
Data.frame
}
\description{
methxy.dmp
}
