% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methxy.vmp.R
\name{methxy.vmp}
\alias{methxy.vmp}
\title{methxy.vmp}
\usage{
methxy.vmp(matrix, design, Sample_Group, cores = 10, chunks = 10)
}
\arguments{
\item{matrix}{matrix}

\item{design}{design}

\item{Sample_Group}{Sample_Group}

\item{cores}{cores}

\item{chunks}{chunks}
}
\value{
res
}
\description{
methxy.vmp
}
