% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateCellCounts2_GMset.R
\name{pickCompProbes}
\alias{pickCompProbes}
\title{pickCompProbes}
\source{
https://bioconductor.org/packages/FlowSorted.Blood.EPIC/
}
\usage{
pickCompProbes(
  mSet,
  cellTypes = NULL,
  numProbes = 50,
  compositeCellType = compositeCellType,
  probeSelect = probeSelect
)
}
\arguments{
\item{mSet}{mSet}

\item{cellTypes}{cellTypes}

\item{numProbes}{numProbes}

\item{compositeCellType}{compositeCellType}

\item{probeSelect}{probeSelect}
}
\value{
res
}
\description{
pickCompProbes
}
\author{
Original authors of FlowSorted.Blood.EPIC
}
