% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.estimate.R
\name{preprocess.estimate}
\alias{preprocess.estimate}
\title{preprocess.estimate}
\usage{
preprocess.estimate(
  DNAm_env,
  save_dir = "./",
  cell.proportion = F,
  CellType = "Blood",
  Referenceset = NULL,
  predict.smok = F,
  predict.age = F
)
}
\arguments{
\item{DNAm_env}{returned by preprocess.filter}

\item{save_dir}{Generates and saves phenotype data using UniqueID
as the filename.}

\item{cell.proportion}{"Whether to perform cell composition estimation.
The default value is 'F'}

\item{CellType}{"Blood" or "Prefrontal_cortex".
Specify which tissue type to use for estimation.
The default value is 'Blood'.}

\item{Referenceset}{To predict cell composition in blood tissue,
the path to the reference file must be provided.}

\item{predict.smok}{The default value is 'F'.Whether to predict smoking score.}

\item{predict.age}{The default value is 'F'.Whether to predict biological age.}
}
\value{
The predicted information has been
added to the target object in the environment.
}
\description{
preprocess.estimate
}
