% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.filter.R
\name{preprocess.filter}
\alias{preprocess.filter}
\title{preprocess.filter}
\usage{
preprocess.filter(DNAm_env)
}
\arguments{
\item{DNAm_env}{The object returned by read_from_idat or read_from_intensity}
}
\value{
An environment containing:
a GenomicMethylSet without Y chromosome ,
a GenomicMethylSet with only Y chromosome ,
the targets file, a UniqueID,
and a message object containing filtering information.
}
\description{
preprocess.filter
}
