% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.norm.R
\name{preprocess.norm}
\alias{preprocess.norm}
\title{preprocess.norm}
\usage{
preprocess.norm(
  DNAm_env,
  Norm = "preprocessQuantile",
  save_result_locally = FALSE,
  save_dir = "./",
  return_res = TRUE
)
}
\arguments{
\item{DNAm_env}{Output returned by preprocess.filter}

\item{Norm}{"preprocessQuantile"}

\item{save_result_locally}{Save results to local directory; default is FALSE}

\item{save_dir}{Directory to save results;default: "./"}

\item{return_res}{Return results; default is TRUE}
}
\value{
The environment contains:
GRSet without sex chromosomes,GRSet with X ,GRSet with Y ,
and the targets table
}
\description{
preprocess.norm
}
