% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_idat.R
\name{read_from_idat}
\alias{read_from_idat}
\title{read_from_idat}
\usage{
read_from_idat(targets_dir, idat_dir, UniqueID, pattern = "GM_ID_Position")
}
\arguments{
\item{targets_dir}{Location of the targets file}

\item{idat_dir}{Location of the idat}

\item{UniqueID}{UniqueID}

\item{pattern}{"Name_ID_Position" or "ID_Position"}
}
\value{
rgset in env
}
\description{
read_from_idat
}
