% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_intensity.R
\name{read_from_intensity}
\alias{read_from_intensity}
\title{read_from_intensity}
\usage{
read_from_intensity(
  targets_dir,
  DNAmType,
  UniqueID,
  pattern = NULL,
  detp = NULL,
  methyl = NULL,
  unmethyl = NULL,
  merge = NULL,
  nThread = 12
)
}
\arguments{
\item{targets_dir}{argets_dir}

\item{DNAmType}{DNAmType}

\item{UniqueID}{UniqueID}

\item{pattern}{1 or 3}

\item{detp}{file_dir}

\item{methyl}{file_dir}

\item{unmethyl}{file_dir}

\item{merge}{file_dir}

\item{nThread}{12}
}
\value{
env
}
\description{
read_from_intensity
}
