% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateCellCounts2_GMset.R
\name{validationCellType}
\alias{validationCellType}
\title{validationCellType}
\source{
https://bioconductor.org/packages/FlowSorted.Blood.EPIC/
}
\usage{
validationCellType(
  Y,
  pheno,
  modelFix,
  modelBatch = NULL,
  L.forFstat = NULL,
  verbose = FALSE
)
}
\arguments{
\item{Y}{Y}

\item{pheno}{pheno}

\item{modelFix}{modelFix}

\item{modelBatch}{modelBatch}

\item{L.forFstat}{L.forFstat}

\item{verbose}{verbose}
}
\value{
res
}
\description{
validationCellType
}
\author{
Original authors of FlowSorted.Blood.EPIC
}
