% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methxy.vmp.R
\name{VarCpG}
\alias{VarCpG}
\title{VarCpG}
\usage{
VarCpG(row, Sample_Group, Probe_name)
}
\arguments{
\item{row}{row}

\item{Sample_Group}{Sample_Group}

\item{Probe_name}{Probe_name}
}
\value{
res
}
\description{
VarCpG
}
